+++ dolphinscheduler-python/pydolphinscheduler/src/pydolphinscheduler/models/queue.py
@@ -15,28 +15,28 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""DolphinScheduler Project object."""
+"""DolphinScheduler User object."""
 
 from typing import Optional
 
-from pydolphinscheduler.core import configuration
-from pydolphinscheduler.core.base_side import BaseSide
-from pydolphinscheduler.java_gateway import launch_gateway
+from pydolphinscheduler import configuration
+from pydolphinscheduler.java_gateway import gateway_result_checker, launch_gateway
+from pydolphinscheduler.models import BaseSide
 
 
-class Project(BaseSide):
-    """DolphinScheduler Project object."""
+class Queue(BaseSide):
+    """DolphinScheduler Queue object."""
 
     def __init__(
         self,
-        name: str = configuration.WORKFLOW_PROJECT,
-        description: Optional[str] = None,
+        name: str = configuration.WORKFLOW_QUEUE,
+        description: Optional[str] = "",
     ):
         super().__init__(name, description)
 
     def create_if_not_exists(self, user=configuration.USER_NAME) -> None:
-        """Create Project if not exists."""
+        """Create Queue if not exists."""
         gateway = launch_gateway()
-        gateway.entry_point.createOrGrantProject(user, self.name, self.description)
-        # TODO recover result checker
-        # gateway_result_checker(result, None)
+        # Here we set Queue.name and Queue.queueName same as self.name
+        result = gateway.entry_point.createProject(user, self.name, self.name)
+        gateway_result_checker(result, None)
