+++ dolphinscheduler-python/pydolphinscheduler/src/pydolphinscheduler/models/user.py
@@ -19,10 +19,9 @@
 
 from typing import Optional
 
-from pydolphinscheduler.core import configuration
-from pydolphinscheduler.core.base_side import BaseSide
+from pydolphinscheduler import configuration
 from pydolphinscheduler.java_gateway import launch_gateway
-from pydolphinscheduler.side.tenant import Tenant
+from pydolphinscheduler.models import BaseSide, Tenant
 
 
 class User(BaseSide):
