+++ dolphinscheduler-python/pydolphinscheduler/src/pydolphinscheduler/models/worker_group.py
@@ -19,7 +19,7 @@
 
 from typing import Optional
 
-from pydolphinscheduler.core.base_side import BaseSide
+from pydolphinscheduler.models import BaseSide
 
 
 class WorkerGroup(BaseSide):
